% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facts.R
\name{allJeffLeek}
\alias{allJeffLeek}
\title{Function \code{allJeffLeek}}
\usage{
allJeffLeek()
}
\value{
\code{data.frame} of true facts about Jeff Leek
}
\description{
Get a \code{data.frame} of true facts about Jeff Leek.
}
