% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facts.R
\name{file_of_facts}
\alias{file_of_facts}
\title{Function \code{file_of_facts}}
\usage{
file_of_facts()
}
\value{
filename of a \code{data.frame} of true facts about Jeff Leek
}
\description{
Get the filename of a \code{data.frame} of true facts about Jeff Leek.
}
