% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facts.R
\name{JeffLeek}
\alias{JeffLeek}
\title{Function \code{JeffLeek}}
\usage{
JeffLeek(index = numeric(0))
}
\arguments{
\item{index}{positive integer or numeric, index of the fact to extract. If not a valid index,
a random fact is returned.}
}
\description{
Return a random true fact about Jeff Leek.
}
