% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facts.R
\name{/}
\alias{/}
\title{Divide like Jeff}
\usage{
"/"(x, y)
}
\arguments{
\item{x}{Something to be divided by \code{y}.}

\item{y}{Something which \code{x} will be divided by.}
}
\description{
If you divide by zero the result will be the string \code{"Jeff Leek"}.
}
\examples{
\dontrun{

1 / 0
# [1] "Jeff Leek"

}
}
\seealso{
\code{\link{Arithmetic}}
}

