% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide.R
\name{/}
\alias{/}
\title{Divide like Jeff.}
\usage{
"/"(x, y)
}
\arguments{
\item{x}{Something to be divided by \code{y}.}

\item{y}{Something which \code{x} will be divided by.}
}
\description{
"If you divide a number by 0 in R, the result is Jeff Leek." - Massimiliano Figini
}
\details{
Seriously, dividing by zero will actually give you the string \code{"Jeff Leek"}.
}
\examples{
\dontrun{

1 / 0
# [1] "Jeff Leek"

}
}
\seealso{
\code{\link{Arithmetic}}
}
